in updateJob and deleteJob checked if the emplyer is the one who created the job to be able to update or delete it

+++ controllers/job.controller.js
@@ -3,30 +3,36 @@ import companyModel from '../models/company.model.js';
 
 // Create a new job
 export const createJob = async (req, res) => {
+  const { id, role } = req.user;
 
-  const {role} = req.user
-
-  if (role !== "employer") {
-    return res.json({message: "You are not allowed to create Job"})
+  if (role !== 'employer') {
+    return res.json({ message: 'You are not allowed to create Job' });
   }
 
   try {
-    const { title, description, salary_range, location, company } = req.body;
+    const { title, description, salary_range, location } = req.body;
 
-    const companyExists = await companyModel.findById(company);
-    if (!companyExists) {
+    const company = await companyModel.findOne({ createdBy: id });
+    if (!company) {
       return res.status(404).json({
         success: false,
-        message: "Company not found",
-      })
-    };
+        message: "user doesn't have a company",
+      });
+    }
+
+    console.log(company);
 
     // Check for required fields
     if (!title || !description || !salary_range || !location) {
-      return res.status(400).json({ message: 'Please provide Title, Description, Salary Range, and Location' });
+      return res.status(400).json({
+        message:
+          'Please provide Title, Description, Salary Range, and Location',
+      });
     }
     if (salary_range.min === undefined || salary_range.max === undefined) {
-      return res.status(400).json({ message: 'Salary range must include both min and max values' });
+      return res
+        .status(400)
+        .json({ message: 'Salary range must include both min and max values' });
     }
 
     // Validate field types
@@ -39,31 +45,44 @@ export const createJob = async (req, res) => {
     if (typeof location !== 'string') {
       return res.status(400).json({ message: 'Location must be a string' });
     }
-    if (typeof salary_range.min !== 'number' || typeof salary_range.max !== 'number') {
-      return res.status(400).json({ message: 'Salary range values must be numbers' });
+    if (
+      typeof salary_range.min !== 'number' ||
+      typeof salary_range.max !== 'number'
+    ) {
+      return res
+        .status(400)
+        .json({ message: 'Salary range values must be numbers' });
     }
 
     // Validate salary_range values
     if (salary_range.min < 0 || salary_range.max < 0) {
-      return res.status(400).json({ message: 'Salary range values cannot be negative' });
+      return res
+        .status(400)
+        .json({ message: 'Salary range values cannot be negative' });
     }
     if (salary_range.min > salary_range.max) {
-      return res.status(400).json({ message: 'Minimum salary must be less than maximum salary' });
+      return res
+        .status(400)
+        .json({ message: 'Minimum salary must be less than maximum salary' });
     }
 
-    const newJob = new jobModel({ 
+    const newJob = new jobModel({
       title,
       description,
       salary_range,
       location,
-      company
+      company: company._id,
       // employer middleware
     });
+
+    console.log(newJob);
 
     const savedJob = await newJob.save();
     res.status(201).json(savedJob);
   } catch (error) {
-    res.status(500).json({ message: 'Sorry! Job is not created', error: error.message });
+    res
+      .status(500)
+      .json({ message: 'Sorry! Job is not created', error: error.message });
   }
 };
 
@@ -73,7 +92,10 @@ export const getAllJobs = async (req, res) => {
     const jobs = await jobModel.find();
     res.status(200).json(jobs);
   } catch (error) {
-    res.status(500).json({ message: 'There are no Jobs at this time', error: error.message });
+    res.status(500).json({
+      message: 'There are no Jobs at this time',
+      error: error.message,
+    });
   }
 };
 
@@ -86,23 +108,32 @@ export const getJobById = async (req, res) => {
     }
     res.status(200).json(job);
   } catch (error) {
-    res.status(500).json({ message: 'There is no job with this id', error: error.message });
+    res
+      .status(500)
+      .json({ message: 'There is no job with this id', error: error.message });
   }
 };
 
 // Update job
 export const updateJob = async (req, res) => {
+  const { id, role } = req.user;
 
-  const {role} = req.user
-
-  if (role !== "employer") {
-    return res.json({message: "You are not allowed to update Job"})
+  if (role !== 'employer') {
+    return res.json({ message: 'You are not allowed to update Job' });
   }
 
   try {
     const job = await jobModel.findById(req.params.id);
     if (!job) {
       return res.status(404).json({ message: 'Job not found' });
+    }
+
+    const company = await companyModel.findById(job.company);
+
+    if (id !== company.createdBy.toString()) {
+      return res.json({
+        message: 'you cannot update a job you did not create',
+      });
     }
 
     const { title, description, salary_range, location } = req.body;
@@ -117,7 +148,9 @@ export const updateJob = async (req, res) => {
     }
     if (description) {
       if (typeof description !== 'string') {
-        return res.status(400).json({ message: 'Description must be a string' });
+        return res
+          .status(400)
+          .json({ message: 'Description must be a string' });
       }
       updatedData.description = description;
     }
@@ -129,38 +162,55 @@ export const updateJob = async (req, res) => {
     }
     if (salary_range) {
       if (salary_range.min === undefined || salary_range.max === undefined) {
-        return res.status(400).json({ message: 'Salary range must include both min and max values' });
+        return res.status(400).json({
+          message: 'Salary range must include both min and max values',
+        });
       }
-      if (typeof salary_range.min !== 'number' || typeof salary_range.max !== 'number') {
-        return res.status(400).json({ message: 'Salary range values must be numbers' });
+      if (
+        typeof salary_range.min !== 'number' ||
+        typeof salary_range.max !== 'number'
+      ) {
+        return res
+          .status(400)
+          .json({ message: 'Salary range values must be numbers' });
       }
       if (salary_range.min < 0 || salary_range.max < 0) {
-        return res.status(400).json({ message: 'Salary range values cannot be negative' });
+        return res
+          .status(400)
+          .json({ message: 'Salary range values cannot be negative' });
       }
       if (salary_range.min > salary_range.max) {
-        return res.status(400).json({ message: 'Minimum salary must be less than maximum salary' });
+        return res
+          .status(400)
+          .json({ message: 'Minimum salary must be less than maximum salary' });
       }
       updatedData.salary_range = salary_range;
     }
 
-    const updatedJob = await jobModel.findByIdAndUpdate(req.params.id, updatedData, {
-      new: true,
-      runValidators: true,
-    });
+    const updatedJob = await jobModel.findByIdAndUpdate(
+      req.params.id,
+      updatedData,
+      {
+        new: true,
+        runValidators: true,
+      }
+    );
 
     res.status(200).json(updatedJob);
   } catch (error) {
-    res.status(500).json({ message: 'An error occurred while updating the job', error: error.message });
+    res.status(500).json({
+      message: 'An error occurred while updating the job',
+      error: error.message,
+    });
   }
 };
 
 // Delete job
 export const deleteJob = async (req, res) => {
+  const { id, role } = req.user;
 
-  const {role} = req.user
-
-  if (role !== "employer") {
-    return res.json({message: "You are not allowed to delete Job"})
+  if (role !== 'employer') {
+    return res.json({ message: 'You are not allowed to delete Job' });
   }
 
   try {
@@ -169,39 +219,48 @@ export const deleteJob = async (req, res) => {
       return res.status(404).json({ message: 'Job not found' });
     }
 
+    const company = await companyModel.findById(job.company);
+
+    if (id !== company.createdBy.toString()) {
+      return res.json({
+        message: 'you cannot delete a job you did not create',
+      });
+    }
+
     await jobModel.findByIdAndDelete(req.params.id);
     res.status(200).json({ message: 'Job deleted successfully' });
   } catch (error) {
-    res.status(500).json({ message: 'Job does not exist!', error: error.message });
+    res
+      .status(500)
+      .json({ message: 'Job does not exist!', error: error.message });
   }
 };
 
 //delete all jobs
 export const deleteAllJobs = async (req, res) => {
+  const { role } = req.user;
 
-  const {role} = req.user
-
-  if (role !== "admin") {
-    return res.json({message: "You are not allowed to delete Jobs"})
+  if (role !== 'admin') {
+    return res.json({ message: 'You are not allowed to delete Jobs' });
   }
 
   try {
     const result = await jobModel.deleteMany({});
 
     if (result.deletedCount > 0) {
-      res.status(200).json({ 
-        message: 'All jobs deleted successfully', 
-        deletedCount: result.deletedCount 
+      res.status(200).json({
+        message: 'All jobs deleted successfully',
+        deletedCount: result.deletedCount,
       });
-    }else {
+    } else {
       res.status(404).json({
-        message: 'No jobs found in the database to delete'
-      })
+        message: 'No jobs found in the database to delete',
+      });
     }
   } catch (error) {
-    res.status(500).json({ 
-      message: 'An error occurred while deleting all jobs', 
-      error: error.message 
+    res.status(500).json({
+      message: 'An error occurred while deleting all jobs',
+      error: error.message,
     });
   }
 };
